requirements.txt
asgiref==3.8.1
attrs==25.1.0
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
